python-iac/tools/jenkins/bin/requirements.txt
boto3
pyyaml
docker
